USe gamescope-plus-git
This probably will fix the Mangohud not displaying in the correct position. Seems like the non-git version uses `wlroots` from the Arch repos while -git version uses a submodue with `wlroots` with a specific commit. It doesn't affect (visibly) on regular gamescope, but since we are rotating the screen on some handhelds it most certainly affect.