Generate some locales (#432)
Generate common locales for the image. This doesn't increase too much in size and will allow for different languages to be selected if the user wants. The list is somewhat arbitrary and "random" but should cover most common languages:
 - English
 - Spanish
 - French
 - German
 - Italian
 - Portuguese
 - Japanese
 - Chinese

More can be added easily but I would stick with the ones Steam client actually supports.