Build our own PKGBUILDs
This will make any PKGBUILD inside `pkgs` directory to be built as a
package and installed in the image like all other AUR packages.
This can be usefu for distro-specific packages or versions of them that
need to be patched.

For now the packages are kept in our repository but eventually we can
add git submodules to keep them in their own repos and even freeze them
to a specific version/build (for more reproducible-ish builds).