Parallel pkg builds (#496)
* split out building docker image

* Implement freeze script
This ensures the docker container gets build with the archive date if required. Only run reflector when it is empty. Putting it in a script ensures it is only one docker layer.

* Implement own-pkg
This builds the own packages in the docker environment and commits the changes. This image can then be used to build all AUR packages.

* implement build-aur stage
Uses pikaur to install package in parallel github actions jobs

* Rebase, and fix review comments

---------

Co-authored-by: alkazar <alkazar@users.noreply.github.com>