Add antimicrox to map gamepad to kbm inputs
This utility should prove very useful for mapping any gamepad to control the keyboard and mouse events when using the desktop. It supports profiles that can be created so any gamepad should easily be configurable including external controllers.

More importantly this utility supports Wayland which up to this day the Steam Input solution does not support.